AULA OOP 4 - CadastroPacientes-Menu Ajuste

+++ AulaOOP3/Revisao/SlnRevisaoOOP/src/Devs2Blu.ProjetosAula.OOP3.Main/Cadastros/CadastroPaciente.cs
@@ -39,7 +39,7 @@ namespace Devs2Blu.ProjetosAula.OOP3.Main.Cadastros
                         break;
                 }
 
-            } while (!opcao.Equals(MenuEnums.SAIR));
+            } while (!opcao.Equals((int)MenuEnums.SAIR));
         }
 
         public void ListarPacientes()
